Change add cmd to add to pasteboard instead
