--- Sources/swift-library/Run.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-func run(cmd: String, args: [String]) {
-    let task = Process()
-    task.launchPath = cmd
-    task.arguments = args
-    task.launch()
-}
