+++ Sources/swift-library/Util.swift
@@ -1,5 +1,6 @@
-import SwiftLibrary
+import Foundation
 import Dispatch
+import SwiftLibrary
 
 func allPackages(query: String) -> [PackageData] {
     let semaphore = DispatchSemaphore(value: 0)
@@ -24,3 +25,22 @@ func allPackages(query: String) -> [PackageData] {
 func firstPackage(query: String) -> PackageData? {
     return allPackages(query: query).first
 }
+
+func run(cmd: String, args: [String]) {
+    let task = Process()
+    task.launchPath = cmd
+    task.arguments = args
+    task.launch()
+}
+
+#if canImport(AppKit)
+import AppKit
+func addToPasteboard(string: String) {
+
+    let pb = NSPasteboard.general
+    pb.string(forType: .string)
+    pb.declareTypes([NSPasteboard.PasteboardType.string], owner: nil)
+    pb.setString(string, forType: .string)
+
+}
+#endif
