+++ Sources/swift-library/main.swift
@@ -36,4 +36,17 @@ case .home(let input):
         exit(1)
     }
     run(cmd: "/usr/bin/open", args: [package.url.absoluteString])
+case .add(let input):
+    guard let package = firstPackage(query: input.package) else {
+        print("No package matching query found.")
+        exit(1)
+    }
+    let dependencyString = ".package(url: \"\(package.url.absoluteString)\", \(input.requirement?.packageString ?? ".branch(\"master\")"))"
+    #if os(macOS)
+    addToPasteboard(string: dependencyString)
+    print("Your clipboard has been updated, just add it to your package manifest.")
+    #else
+    print("Copy the following and add it to your package manifest.")
+    print(dependencyString)
+    #endif
 }
